+++ src/modules/CanvasRotate/CanvasRotate.hpp
@@ -14,6 +14,7 @@ class $editorModule(CanvasRotate) {
 
     void onEditor() override;
     void toggleBetterEditHook(bool enabled);
+    bool isLassoActive();
 
     bool onToggled(bool state) override;
     bool onSettingChanged(std::string_view key, const matjson::Value& value) override;