+++ src/modules/StartPosTools/StartPosTools.cpp
@@ -7,6 +7,28 @@ void StartPosTools::onEditor() {
     m_overlay = StartPosOverlay::create();
 	m_overlay->setID("start-pos-controls"_spr);
 	m_editorLayer->m_objectLayer->addChild(m_overlay);
+
+    auto editorUI = static_cast<SPTEditorUI*>(m_editorUI);
+    auto fields = editorUI->m_fields.self();
+
+    auto spr = CCSprite::create("playtest-start-pos.png"_spr);
+    spr->setScale(0.75f);
+    fields->m_startPosBtn = CCMenuItemSpriteExtra::create(spr, m_editorUI, menu_selector(SPTEditorUI::onStartPosPlaytest));
+    fields->m_startPosBtn->setTag(1);
+    editorUI->m_uiItems->addObject(fields->m_startPosBtn);
+
+    auto playtestMenu = editorUI->getChildByID("playtest-menu");
+    if (playtestMenu) {
+        auto layout = static_cast<AxisLayout*>(playtestMenu->getLayout());
+        if (layout) {
+            layout->ignoreInvisibleChildren(true);
+            layout->setAutoScale(false);
+            layout->setAutoGrowAxis(0);
+        }
+        playtestMenu->addChild(fields->m_startPosBtn);
+        playtestMenu->updateLayout();
+        editorUI->updatePlaytestMenu();
+    }
 }
 
 void StartPosTools::updateOverlay() {
@@ -17,6 +39,46 @@ void StartPosTools::updateOverlay() {
     if (auto startPos = typeinfo_cast<StartPosObject*>(m_editorUI->m_selectedObject)) {
         m_overlay->setStartPos(startPos);
     }
+}
+
+void SPTEditorUI::updatePlaytestMenu() {
+    runAction(CallFuncExt::create([this] {
+        auto playtestMenu = getChildByID("playtest-menu");
+        if (playtestMenu) {
+            playtestMenu->updateLayout();
+            auto playbackMenu = getChildByID("playback-menu");
+            if (playbackMenu) {
+                playtestMenu->setPositionX(playbackMenu->getPositionX());
+            }
+            auto objectInfoLabel = getChildByID("object-info-label");
+            if (objectInfoLabel) {
+                objectInfoLabel->setPositionX(playtestMenu->boundingBox().getMaxX() + 5);
+            }
+        }
+    }));
+}
+
+void SPTEditorUI::showUI(bool show) {
+    EditorUI::showUI(show);
+    auto fields = m_fields.self();
+    if (show && m_editorLayer->m_playbackMode == PlaybackMode::Paused) {
+        fields->m_startPosBtn->setVisible(false);
+    }
+    updatePlaytestMenu();
+}
+
+void SPTEditorUI::onPlaytest(cocos2d::CCObject* sender) {
+    auto fields = m_fields.self();
+    if (sender->getTag() != 1) {
+        fields->m_fromStart = true;
+    }
+    EditorUI::onPlaytest(sender);
+    fields->m_fromStart = false;
+    updatePlaytestMenu();
+}
+
+void SPTEditorUI::onStartPosPlaytest(CCObject* sender) {
+    EditorUI::get()->onPlaytest(sender);
 }
 
 void SPTEditorUI::deselectAll() {
@@ -48,9 +110,19 @@ bool SPTEditorUI::shouldDeleteObject(GameObject* object) {
 }
 
 void SPTGJBaseGameLayer::orderSpawnObjects() {
-    auto startPos = StartPosTools::get()->m_overlay->getLastStartPos();
-    if (startPos) {
-        LevelEditorLayer::get()->setStartPosObject(startPos);
+    auto editorUI = static_cast<SPTEditorUI*>(EditorUI::get());
+    if (!editorUI) return GJBaseGameLayer::orderSpawnObjects();
+
+    auto fields = editorUI->m_fields.self();
+    if (fields->m_fromStart) {
+        LevelEditorLayer::get()->setStartPosObject(nullptr);
     }
+    else {
+        auto startPos = StartPosTools::get()->m_overlay->getLastStartPos();
+        if (startPos) {
+            LevelEditorLayer::get()->setStartPosObject(startPos);
+        }
+    }
+    
     GJBaseGameLayer::orderSpawnObjects();
 }